backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1

# S3/R2 compatibility
urllib3<2.0.0
django-storages>=1.14.2
boto3>=1.35.0
botocore>=1.35.0
Pillow>=10.2.0

# Email and SMS
sendgrid==6.11.0
twilio==8.12.0

# Calendar integration
# Calendar integration
caldav==1.3.9
icalendar==5.0.11

# Payments
stripe>=5.0.0

# Utilities

python-magic==0.4.27
bleach==6.1.0
markdown==3.5.2
pytz==2024.1

# Dev tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-jazzmin==2.6.0
icalendar==5.0.11
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
